# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSI",
        "version": "1.0"
    },
    "paths": {
...
        "/api/salle/{id}": {
            "get": {
                "tags": [
                    "Salle"
                ],
                "description": "Cette route permet de retourner une salle en fonction de son id au format Json.",
                "operationId": "caa70e685ed9d37a804ac82f3876765a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de la salle que nous voulons récupérer !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne une salle au format JSON.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Salle"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/edit-salle/{id}": {
            "put": {
                "tags": [
                    "Salle"
                ],
                "description": "Cette route permet de modifier une salle en base de données.",
                "operationId": "55fc3b62eac4a9aa01147df696297e85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de la salle que nous voulons modifier !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne la salle modifiée au format JSON.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Salle"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete-salle/{id}": {
            "delete": {
                "tags": [
                    "Salle"
                ],
                "description": "Cette route permet de supprimer une salle en base de données.",
                "operationId": "c00d03611e3dc6220d98e31858d12e80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de la salle que nous voulons supprimer !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne un message de succès si la suppression du ticket a eu lieu en base de données."
                    },
                    "500": {
                        "description": "Retourne un message d'erreur si la suppression du ticket n'a pas eu lieu en base de données."
                    }
                }
            }
        },
...
        "/api/utilisateur/{email}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Cette route permet de retourner un utilisateur en fonction de son email au format Json.",
                "operationId": "9abcba0b3067cb5ac561dfa4d6f9b7a9",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "L'email de l'utilisateur que nous voulons récupérer !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne un message de succès si la suppression du ticket a eu lieu en base de données."
                    },
                    "500": {
                        "description": "Retourne un message d'erreur si la suppression du ticket n'a pas eu lieu en base de données."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Etat": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "libelle": {
                        "type": "string"
                    },
                    "ordre": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Salle": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "libelle": {
                        "type": "string"
                    },
                    "tickets": {
                        "$ref": "#/components/schemas/Ticket"
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "mail_expediteur": {
                        "type": "string"
                    },
                    "mail_destinataire": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "objet": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "date_creation": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "date_limite": {
                        "type": "string",
                        "format": "date"
                    },
                    "etat": {
                        "$ref": "#/components/schemas/Etat"
                    },
                    "commentaire": {
                        "type": "App\\Entity\\Commentaire",
                        "format": "ArrayObject"
                    },
                    "utilisateur": {
                        "type": "App\\Entity\\User",
                        "format": "App\\Entity\\ArrayObject"
                    }
                },
                "type": "object"
            },
...
        }
    }
}